% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acs_conf}
\alias{acs_conf}
\title{American Community Survey confidential microdata (with weights)}
\format{
\subsection{\code{acs_conf}}{

A data frame with 1,500 rows and 12 columns:
\describe{
\item{county}{fct, county}
\item{gq}{fct, group quarter kind}
\item{sex}{fct, sex}
\item{marst}{fct, marital status}
\item{hcovany}{fct, health insurance status}
\item{empstat}{fct, employment status; contains empty levels.}
\item{classwkr}{fct, employment kind (ex: self-employed, etc.);
contains "N/A" levels.}
\item{age}{dbl, age (in years)}
\item{famsize}{dbl, household/family size}
\item{transit_time}{dbl, transit time to work (in minutes)}
\item{inctot}{dbl, annual income; contains missing values}
\item{wgt}{dbl, survey weight}
}
}
}
\source{
\url{https://usa.ipums.org/usa/}
}
\usage{
acs_conf
}
\description{
An extract constructed from the 2019 American Community Survey containing a
survey sample of n = 1500 Nebraska respondents, with survey weights included.
}
\details{
Original data source:
Steven Ruggles, Sarah Flood, Matthew Sobek, Daniel Backman, Annie Chen,
Grace Cooper, Stephanie Richards, Renae Rogers, and Megan Schouweiler.
IPUMS USA: Version 15.0 [dataset]. Minneapolis, MN: IPUMS, 2024.
https://doi.org/10.18128/D010.V15.0
}
\keyword{datasets}
