% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acs_start_nw}
\alias{acs_start_nw}
\title{American Community Survey starting microdata (without weights)}
\format{
\subsection{\code{acs_start_nw}}{

A data frame with 500 rows and 4 columns:
\describe{
\item{county}{fct, county}
\item{gq}{fct, group quarter kind}
\item{sex}{fct, sex}
\item{marst}{fct, marital status}
}
}
}
\source{
\url{https://usa.ipums.org/usa/}
}
\usage{
acs_start_nw
}
\description{
An extract constructed from the 2019 American Community Survey containing a
survey sample of n = 500 Nebraska respondents, with survey weights included.
}
\details{
Original data source:
Steven Ruggles, Sarah Flood, Matthew Sobek, Daniel Backman, Annie Chen,
Grace Cooper, Stephanie Richards, Renae Rogers, and Megan Schouweiler.
IPUMS USA: Version 15.0 [dataset]. Minneapolis, MN: IPUMS, 2024.
https://doi.org/10.18128/D010.V15.0
}
\keyword{datasets}
