% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visit_sequence_api.R
\name{add_sequence_manual}
\alias{add_sequence_manual}
\title{Add to visit sequence using a manual method}
\usage{
add_sequence_manual(roadmap, ...)
}
\arguments{
\item{roadmap}{A \code{roadmap} object.}

\item{...}{\if{html}{\out{<tidy-select>}} One or more unquoted expressions separated by
commas. Variable names can be used as if they were positions in the data
frame, so expressions like x:y can be used to select a range of variables.}
}
\value{
An updated \code{roadmap} object.
}
\description{
Add to visit sequence using a manual method
}
\examples{

roadmap(
  conf_data = acs_conf_nw, 
  start_data = acs_start_nw
) |>
  add_sequence_manual(
    c("inctot", "hcovany", "empstat",  "classwkr",  "age",   "famsize",  
      "transit_time")
  )

}
