% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_constraints.R
\name{assign_constraints_cat}
\alias{assign_constraints_cat}
\title{Assign constraints to a synthetic data frame for categorical variables.}
\usage{
assign_constraints_cat(synth_data, constraints)
}
\arguments{
\item{synth_data}{A synthetic data set}

\item{constraints}{A list of constraints created by the \code{constraints}
constructor.}
}
\value{
A list of two data.frames: \code{synth_data} and \code{condition_defs}
}
\description{
This function serves two functions. First, it maps each row in the synthetic
data to an integer \code{.condition_id} that determines how to enforce categorical
constraints. Second, it creates a new dataframe, \code{condition_defs}, that stores
information about how to implement categorical constraints for a specific
\code{.condition_id} value.
}
