% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_constraints.R
\name{assign_constraints_num}
\alias{assign_constraints_num}
\title{Assign constraints to a synthetic data frame for numeric variables}
\usage{
assign_constraints_num(synth_data, constraints)
}
\arguments{
\item{synth_data}{A synthetic data set}

\item{constraints}{A list of constraints created by the \code{constraints}
constructor.}
}
\value{
A synthetic data frame with a \code{min} and \code{max} column added.
}
\description{
Assign constraints to a synthetic data frame for numeric variables
}
