% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_na.R
\name{collapse_na}
\alias{collapse_na}
\title{Collapse data frames with _NA variables to coerce related variables to
include NA}
\usage{
collapse_na(data)
}
\arguments{
\item{data}{A data frame with columns ending in _NA}
}
\value{
A data frame with no _NA columns and NA values
}
\description{
Collapse data frames with _NA variables to coerce related variables to
include NA
}
