% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{constraints_api}
\alias{constraints_api}
\alias{add_constraints}
\alias{update_constraints}
\alias{reset_constraints}
\title{Add, update, or reset a \code{constraints} object within an existing \code{roadmap}.}
\usage{
add_constraints(roadmap, constraints)

update_constraints(roadmap, ...)

reset_constraints(roadmap)
}
\arguments{
\item{roadmap}{A \code{roadmap} object}

\item{constraints}{A \code{constraints} object.}

\item{...}{Optional named parameters passed to \code{constraints()}.}
}
\value{
A new \code{roadmap} object.
}
\description{
Add, update, or reset a \code{constraints} object within an existing \code{roadmap}.
}
