% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_col_schema.R
\name{construct_col_schema}
\alias{construct_col_schema}
\title{Construct a list of \code{col_schema}s that aligns with a visit sequence.}
\usage{
construct_col_schema(roadmap)
}
\arguments{
\item{roadmap}{A \code{roadmap} S3 object}
}
\value{
A named list of lists mapping variable names to their \code{col_schema}
specifications.
}
\description{
Construct a list of \code{col_schema}s that aligns with a visit sequence.
}
