% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_constraints.R
\name{construct_constraints}
\alias{construct_constraints}
\title{Construct a list of constraints that aligns with a visit sequence.}
\usage{
construct_constraints(roadmap)
}
\arguments{
\item{roadmap}{A \code{roadmap} S3 object}
}
\value{
A named list of lists mapping variable names to their \code{constraint}
dataframes and \code{max_z} values
}
\description{
This set of constraints respects the \code{visit_sequence} ordering and properly
respects default and custom settings for different modes of constraint
enforcement.
}
