% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_extractors.R
\name{construct_extractors}
\alias{construct_extractors}
\title{Construct a list of extractors for parsnip models}
\usage{
construct_extractors(
  roadmap,
  default_extractor = NULL,
  custom_extractors = NULL
)
}
\arguments{
\item{roadmap}{A roadmap object}

\item{default_extractor}{An extractor from library(parsnip)}

\item{custom_extractors}{A formatted list of extractors}
}
\value{
A named list of extractors
}
\description{
Construct a list of extractors for parsnip models
}
\examples{

roadmap <- roadmap(
  conf_data = acs_conf_nw,
  start_data = acs_start_nw
)

extractors_default <- construct_extractors(
  roadmap = roadmap, 
  default_extractor = parsnip::extract_fit_engine
)

}
