% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_noise_kde.R
\name{.calc_bandwidths}
\alias{.calc_bandwidths}
\title{Calculate bandwidths for each ntile in a confidential vector}
\usage{
.calc_bandwidths(baseline, n, ties_method = "collapse")
}
\arguments{
\item{baseline}{A numeric vector of confidential data}

\item{n}{An integer for the number of ntiles}

\item{ties_method}{The ntiles approach to adding noise requuires a one-to-one
mapping from model-generated values to ntiles in the original data. The
methods "collapse", "random", and "exclusions" deal with situations where the
ntiles lack unique bounds.}
}
\value{
A data frame with ntiles and bandwidths
}
\description{
Calculate bandwidths for each ntile in a confidential vector
}
\examples{

.calc_bandwidths(baseline = mtcars$mpg, n = 2)

}
