% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presynth.R
\name{.construct_workflows}
\alias{.construct_workflows}
\title{Construct tidysynthesis workflows from a \code{roadmap} and \code{synth_spec}}
\usage{
.construct_workflows(roadmap, synth_spec)
}
\arguments{
\item{roadmap}{A \code{roadmap} object}

\item{synth_spec}{A \code{synth_spec} object}
}
\value{
A list of built models, recipes, samplers, etc.
}
\description{
Construct tidysynthesis workflows from a \code{roadmap} and \code{synth_spec}
}
