% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_noise_kde.R
\name{.create_ntiles}
\alias{.create_ntiles}
\title{Calculate ntiles on confidential data and apply to predicted values}
\usage{
.create_ntiles(
  x_bounds,
  y_ntiles,
  n,
  ties_method = c("collapse", "random", "exclusions"),
  na.rm = FALSE
)
}
\arguments{
\item{x_bounds}{A numeric vector used to calculate the ntile bounds}

\item{y_ntiles}{A numeric vector to which the bounds from x_bounds are applied}

\item{n}{The number of ntiles to create}

\item{ties_method}{The ntiles approach to adding noise requires a one-to-one
mapping from model-generated values to ntiles in the original data. The
methods "collapse", "random", and "exclusions" deal with situations where the
ntiles lack unique bounds.}

\item{na.rm}{logical; if true, any \code{NA} and \code{NaN}'s are removed from x_bounds
before the quantiles are computed.}
}
\value{
A numeric vector of ntiles
}
\description{
Parts of the x distribution with many repeated values can lead to duplicated
bounds. Duplicated bounds are combined into one ntile for the y distribution.
}
