% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_utils.R
\name{.extract_mode}
\alias{.extract_mode}
\title{Extract named vector of model modes from a roadmap}
\usage{
.extract_mode(roadmap)
}
\arguments{
\item{roadmap}{A \code{roadmap} S3 object.}
}
\value{
A named vector of \code{regression} and \code{classification} strings
}
\description{
Extract named vector of model modes from a roadmap
}
