% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synth_component_utils.R
\name{.identical_funcs}
\alias{.identical_funcs}
\title{Helper for functional equivalent of \code{identical()} for functions}
\usage{
.identical_funcs(x, y)
}
\arguments{
\item{x}{function}

\item{y}{function}
}
\value{
Logical
}
\description{
Helper for functional equivalent of \code{identical()} for functions
}
