% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synth_component_utils.R
\name{.map_model_to_default_sampler}
\alias{.map_model_to_default_sampler}
\title{Map model object to default sampler function}
\usage{
.map_model_to_default_sampler(model)
}
\arguments{
\item{model}{A \code{parsnip::model_spec} object}
}
\value{
A sampler function
}
\description{
Map model object to default sampler function
}
