% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synth_component_utils.R
\name{.name_to_inspect}
\alias{.name_to_inspect}
\title{constant mapping between component names and inspections}
\usage{
.name_to_inspect
}
\description{
constant mapping between component names and inspections
}
