% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synth_component_utils.R
\name{.update_custom_components}
\alias{.update_custom_components}
\title{Update custom_components}
\usage{
.update_custom_components(synth_spec, component_name, custom_name, ...)
}
\arguments{
\item{synth_spec}{A \code{synth_spec} object}

\item{component_name}{String, one of the names in \code{names(.name_to_inspect)}}

\item{custom_name}{String, name for \verb{custom_<>} components}

\item{...}{Optional named lists with two elements, \code{vars} and \verb{<component_name>}
mapping variable names to their corresponding \code{synth_spec} objects}
}
\description{
Update custom_components
}
