% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_utils.R
\name{.validate_construct_inputs_optional}
\alias{.validate_construct_inputs_optional}
\title{Validate construct_* functions for optional components (for internal use only)}
\usage{
.validate_construct_inputs_optional(
  visit_sequence,
  default_reg,
  default_class,
  custom_list,
  type_check_func,
  obj_name
)
}
\arguments{
\item{visit_sequence}{A character vector of \code{visit_sequence} variable names.}

\item{default_reg}{A default object used for regression models}

\item{default_class}{A default object used for classification models}

\item{custom_list}{A list of named lists mapping variables to objects}

\item{type_check_func}{A function that returns a boolean if the object if
of the correct type.}

\item{obj_name}{A string describing the object name, used only for printing
explanatory error messages.}
}
\description{
Validate construct_* functions for optional components (for internal use only)
}
