% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presynth.R
\name{.validate_presynth}
\alias{.validate_presynth}
\title{\code{presynth} validator}
\usage{
.validate_presynth(presynth)
}
\arguments{
\item{presynth}{A \code{presynth} object}
}
\description{
Ensures compability between \code{synth_spec} and \code{roadmap} objects, raising
an error if not.
}
