% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enforce_na.R
\name{enforce_na}
\alias{enforce_na}
\title{Add missing values where values should be missing according to _NA variables}
\usage{
enforce_na(data)
}
\arguments{
\item{data}{A synthetic data frame with _NA columns}
}
\value{
A synthetic data frame with _NA columns that converts values that
are labelled missing in an _NA variable to missing in the corresponding
variable
}
\description{
Add missing values where values should be missing according to _NA variables
}
\examples{

example_na_expanded <- expand_na(data = example_na)

enforce_na(data = example_na_expanded)

}
