% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_na.R
\docType{data}
\name{example_na}
\alias{example_na}
\title{A df with different types of missingness}
\format{
A tibble with 200 observations
and 6 variables:
\describe{
\item{age}{Age of respondent}
\item{sex}{Sex of respondent with missingness at random}
\item{labor_force}{Labor force status of respondent with structural missingness}
\item{hours}{Hours work of respondent with missingness at random}
\item{wages}{Wages earned with structural missingness}
}
}
\usage{
example_na
}
\description{
A df with different types of missingness
}
\keyword{datasets}
