% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noise.R
\name{exec_noise_func}
\alias{exec_noise_func}
\title{Execute a noise function within \code{generate_predictions()}}
\usage{
exec_noise_func(
  model,
  new_data,
  conf_model_data,
  outcome_var,
  col_schema,
  pred,
  noise
)
}
\arguments{
\item{model}{A \code{model_spec} or a list of \code{model_spec}s from \code{library(parsnip)}}

\item{new_data}{A data frame used to generate predictions}

\item{conf_model_data}{A data frame for estimating the predictive model}

\item{outcome_var}{A string name representing the outcome variable}

\item{col_schema}{A list of column schema specifications for the new variable}

\item{pred}{A vector of values predicted by the model}

\item{noise}{A \code{noise} S3 object}
}
\value{
A vector of noisy predictions with the same length as \code{pred}
}
\description{
Execute a noise function within \code{generate_predictions()}
}
