% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_method.R
\name{exec_start_method}
\alias{exec_start_method}
\title{Execute a \code{start_method} instance}
\usage{
exec_start_method(roadmap)
}
\arguments{
\item{roadmap}{A \code{roadmap} object}
}
\value{
A \code{data.frame}
}
\description{
Execute a \code{start_method} instance
}
