% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_na.R
\name{expand_na}
\alias{expand_na}
\title{Add new variables that indicate if a value is "missing" or "not missing" for
original variables that contain NA}
\usage{
expand_na(
  data,
  types = c("chr", "dbl", "fct", "lgl", "int", "ord"),
  skip_vars = NULL
)
}
\arguments{
\item{data}{A data frame}

\item{types}{A vector of variables types to expand}

\item{skip_vars}{A character vector of variables that shouldn't be expanded}
}
\value{
An augmented data frame with the original variables and new
variables that contain the missingness patterns of variables with NA
}
\description{
Add new variables that indicate if a value is "missing" or "not missing" for
original variables that contain NA
}
