% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_predictions.R
\name{generate_predictions}
\alias{generate_predictions}
\title{wrapper function for generating predictions}
\usage{
generate_predictions(
  model,
  new_data,
  conf_model_data,
  outcome_var,
  col_schema,
  sampler,
  noise,
  constraints_df,
  condition_defs,
  factor_levels,
  max_z,
  ysyn,
  invert_transformations,
  calc_ldiversity,
  store_ldiversity = NULL
)
}
\arguments{
\item{model}{A \code{model_spec} or a list of \code{model_spec}s from \code{library(parsnip)}}

\item{new_data}{A data frame used to generate predictions}

\item{conf_model_data}{A data frame for estimating the predictive model}

\item{outcome_var}{A string name representing the outcome variable}

\item{col_schema}{A list of column schema specifications for the new variable}

\item{sampler}{A method used to predict values from estimated models.}

\item{noise}{A \code{noise} S3 object}

\item{constraints_df}{A data frame of constraints}

\item{condition_defs}{A data.frame of condition definitions}

\item{factor_levels}{A vector of allowable factor levels.}

\item{max_z}{An integer for the number of z-bounding iterations before hard
bounding. A 0 will simply hard bound without iterating predictions}

\item{ysyn}{A vector of predicted values}

\item{invert_transformations}{A Boolean for if outcome variable
transformations should be inverted during synthesis}

\item{calc_ldiversity}{A Boolean for if ldiversity should be calculated}

\item{store_ldiversity}{A vector for storing ldiversity}
}
\value{
A list with synthetic values and calulcated ldiveristy
}
\description{
This function heavily relies on recursion.
}
