% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invert.R
\name{invert}
\alias{invert}
\alias{invert.recipe}
\title{An S3 method for inverting a step}
\usage{
invert(object, predictions, ...)
}
\arguments{
\item{object}{A recipe after fitting a model}

\item{predictions}{A data frame with .pred}

\item{...}{Other arguments}
}
\description{
An S3 method for inverting a step
}
\concept{postprocessing}
