% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invert.R
\name{invert.step_BoxCox}
\alias{invert.step_BoxCox}
\title{Invert a Box-Cox transformation}
\usage{
\method{invert}{step_BoxCox}(object, predictions, ...)
}
\arguments{
\item{object}{A recipe after fitting a model}

\item{predictions}{A data frame with .pred}

\item{...}{Other arguments}
}
\description{
Invert a Box-Cox transformation
}
