% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yeo_johnson.R
\name{is.Numeric}
\alias{is.Numeric}
\title{Ported from library(VGAM)}
\usage{
is.Numeric(x, length.arg = Inf, integer.valued = FALSE, positive = FALSE)
}
\arguments{
\item{x}{A vector}

\item{length.arg}{A numeric}

\item{integer.valued}{A logical}

\item{positive}{A logical}
}
\description{
Ported from library(VGAM)
}
