% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synth_spec.R
\name{is_synth_spec}
\alias{is_synth_spec}
\title{Check if object is \code{synth_spec}}
\usage{
is_synth_spec(z)
}
\arguments{
\item{z}{Object}
}
\value{
Logical
}
\description{
Check if object is \code{synth_spec}
}
