% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presynth.R
\name{new_presynth}
\alias{new_presynth}
\title{\code{presynth} constructor}
\usage{
new_presynth(roadmap, synth_spec)
}
\arguments{
\item{roadmap}{A \code{roadmap} object}

\item{synth_spec}{A \code{synth_spec} object}
}
\description{
\code{presynth} constructor
}
