% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roadmap.R
\name{new_roadmap}
\alias{new_roadmap}
\title{Roadmap constructor}
\usage{
new_roadmap(
  conf_data,
  start_data,
  start_method = NULL,
  schema = NULL,
  visit_sequence = NULL,
  replicates = NULL,
  constraints = NULL
)
}
\arguments{
\item{conf_data}{A \code{data.frame} of confidential data.}

\item{start_data}{A \code{data.frame} of starting data used to initialize the process.}

\item{start_method}{An optional \code{start_method} object.}

\item{schema}{An optional \code{schema} object.}

\item{visit_sequence}{An optional \code{visit_sequence} object.}

\item{replicates}{An optional \code{replicates} object.}

\item{constraints}{An optional \code{constraints} object.}
}
\value{
A new \code{roadmap} object.
}
\description{
Roadmap constructor
}
