% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visit_sequence.R
\name{new_visit_sequence}
\alias{new_visit_sequence}
\title{Constructor for \code{visit_sequence}}
\usage{
new_visit_sequence(schema, weight_var = NULL, synthesize_weight = TRUE)
}
\arguments{
\item{schema}{A \code{schema} object.}

\item{weight_var}{A numeric weight for the weighted total ordering.}

\item{synthesize_weight}{Boolean for if weight_var should be included in the
visit sequence.}
}
\value{
A \code{visit_sequence} object.
}
\description{
Constructor for \code{visit_sequence}
}
