% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noise.R
\name{noise}
\alias{noise}
\title{Create a noise object}
\usage{
noise(add_noise = FALSE, mode = "regression", noise_func = NULL, ...)
}
\arguments{
\item{add_noise}{Boolean, TRUE if adding noise}

\item{mode}{String, one of "regression" or "classification"}

\item{noise_func}{A function that adds noise to}

\item{...}{Optional named additional arguments to pass to \code{noise_func(...)}}
}
\description{
Create a noise object
}
