% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postsynth.R
\name{postsynth}
\alias{postsynth}
\title{Create a postsynth object}
\usage{
postsynth(
  synthetic_data,
  jth_preprocessing,
  total_synthesis_time,
  jth_synthesis_time,
  extractions,
  ldiversity
)
}
\arguments{
\item{synthetic_data}{A synthetic data set.}

\item{jth_preprocessing}{A list of recipes from \code{library(recipe)}.}

\item{total_synthesis_time}{A double with the total time of synthesis.}

\item{jth_synthesis_time}{A vector or dataframe of synthesis times.}

\item{extractions}{A list with extracted output from the synthesis model.}

\item{ldiversity}{A vector or dataframe of ldiversity stats.}
}
\value{
A \code{postsynth} object.
}
\description{
Create a postsynth object
}
