% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset_postsynth.R
\name{postsynth_to_synth_spec}
\alias{postsynth_to_synth_spec}
\title{Convert a incompletely synthesized \code{postsynth} into a new \code{synth_spec} using
the same \code{synth_spec} settings from the incomplete synthesis.}
\usage{
postsynth_to_synth_spec(postsynth)
}
\arguments{
\item{postsynth}{A \code{postsynth} object generated with \code{keep_workflows == TRUE}}
}
\value{
A \code{synth_spec} object that is similar to the initial \code{synth_spec}
but updated to remove information about already-synthesized variables.
}
\description{
Convert a incompletely synthesized \code{postsynth} into a new \code{synth_spec} using
the same \code{synth_spec} settings from the incomplete synthesis.
}
\examples{

# create roadmap
rm <- roadmap(
  conf_data = acs_conf_nw,
  start_data = acs_start_nw
) 

rpart_mod_reg <- parsnip::decision_tree() |>
  parsnip::set_engine(engine = "rpart") |>
  parsnip::set_mode(mode = "regression")

rpart_mod_class <- parsnip::decision_tree() |>
  parsnip::set_engine(engine = "rpart") |>
  parsnip::set_mode(mode = "classification")

synth_spec1 <- synth_spec(
  default_regression_model = rpart_mod_reg,
  default_regression_sampler = sample_rpart,
  default_classification_model = rpart_mod_class,
  default_classification_sampler = sample_rpart
)

# create a presynth object
# use defaults for noise, constraints, and replicates
presynth1 <- presynth(
  roadmap = rm,
  synth_spec = synth_spec1
)

# synthesize with keep_workflows = TRUE
set.seed(1)
postsynth1 <- synthesize(
  presynth = presynth1, 
  keep_workflows = TRUE
)

# return roadmap (will return original roadmap if synthesis completes).
new_spec <- postsynth_to_synth_spec(postsynth1)

}
