% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noise.R
\name{print.noise}
\alias{print.noise}
\title{Print the noise object to the console with formatting}
\usage{
\method{print}{noise}(x, ...)
}
\arguments{
\item{x}{A \code{noise} object}

\item{...}{further arguments passed to or from other methods (not currently
used).}
}
\description{
Print the noise object to the console with formatting
}
\examples{

print(noise())

}
