% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_method.R
\name{print.start_method}
\alias{print.start_method}
\title{Print the start_method object to the console with formatting}
\usage{
\method{print}{start_method}(x, ...)
}
\arguments{
\item{x}{A \code{start_method} object}

\item{...}{further arguments passed to or from other methods (not currently
used).}
}
\value{
A \code{start_method} object
}
\description{
Print the start_method object to the console with formatting
}
\examples{
 

print(start_method())

}
