% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synth_spec.R
\name{print.synth_spec}
\alias{print.synth_spec}
\title{Print the replicates object to the console with formatting}
\usage{
\method{print}{synth_spec}(x, ...)
}
\arguments{
\item{x}{A \code{replicates} object}

\item{...}{further arguments passed to or from other methods (not currently
used).}
}
\value{
A \code{synth_spec} object
}
\description{
Print the replicates object to the console with formatting
}
\examples{

synth_spec <- synth_spec()

print(synth_spec)

}
