% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visit_sequence.R
\name{print.visit_sequence}
\alias{print.visit_sequence}
\title{Print method for \code{visit_sequence} objects}
\usage{
\method{print}{visit_sequence}(x, ...)
}
\arguments{
\item{x}{A \code{visit_sequence} object}

\item{...}{further arguments passed to or from other methods (not currently
used).}
}
\value{
Invisibly returns the input \code{visit_sequence} object.
}
\description{
Print method for \code{visit_sequence} objects
}
\examples{

rm <- roadmap(
  conf_data = acs_conf_nw, 
  start_data = acs_start_nw
)

print(rm[["visit_sequence"]])

}
