% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicates.R
\name{replicates}
\alias{replicates}
\title{Create a replicates object}
\usage{
replicates(
  start_data_replicates = 1,
  model_sample_replicates = 1,
  end_to_end_replicates = 1
)
}
\arguments{
\item{start_data_replicates}{The number of starting data replicates to use.
Note that if no \code{start_method} is provided, all start data replicates will
be identical.}

\item{model_sample_replicates}{The number of replicates for the conditional
modeling process, including modeling and sampling new synthetic values.}

\item{end_to_end_replicates}{The number of replicates for the entire synthesis process,
including all previously specified steps.}
}
\value{
A new \code{replicates} object.
}
\description{
Create a replicates object
}
