% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicates.R
\name{replicates_api}
\alias{replicates_api}
\alias{add_replicates}
\alias{update_replicates}
\alias{reset_replicates}
\title{Add, update, or reset a \code{replicates} object within an existing \code{roadmap}.}
\usage{
add_replicates(roadmap, replicates)

update_replicates(roadmap, ...)

reset_replicates(roadmap)
}
\arguments{
\item{roadmap}{A \code{roadmap} object}

\item{replicates}{A \code{replicates} object.}

\item{...}{Optional named parameters passed to \code{replicates()}.}
}
\value{
A new \code{roadmap} object.
}
\description{
Add, update, or reset a \code{replicates} object within an existing \code{roadmap}.
}
