% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_rpart.R
\name{sample_rpart}
\alias{sample_rpart}
\title{Sample the conditional distribution created by a CART model}
\usage{
sample_rpart(model, new_data, conf_data, ignore_zeros = TRUE)
}
\arguments{
\item{model}{A "model_fit" object created by rpart}

\item{new_data}{A data frame with predictors}

\item{conf_data}{A data frame with original confidential predictors}

\item{ignore_zeros}{Should a vector of all 0 observations return NA for
the l-diversity calculation. Defaults to \code{TRUE}.}
}
\value{
A numeric vector of predictions
}
\description{
Sample the conditional distribution created by a CART model
}
