% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synth_component_utils.R
\name{synth_spec_is_component}
\alias{synth_spec_is_component}
\title{Inspections for \code{synth_spec} components}
\arguments{
\item{z}{Object}
}
\value{
Boolean if matches class type
}
\description{
Inspections for \code{synth_spec} components
}
