% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synth_spec.R
\name{synth_spec_noise_api}
\alias{synth_spec_noise_api}
\title{Add, update, or remove noise from a \code{synth_spec} object}
\arguments{
\item{synth_spec}{A \code{synth_spec} object}

\item{...}{Optional named lists with two elements, \code{vars} and \code{noise},
mapping variable names to samplers.}
}
\value{
A new \code{synth_spec} object.
}
\description{
Add, update, or remove noise from a \code{synth_spec} object
}
