% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synth_spec.R
\name{synth_spec_recipes_api}
\alias{synth_spec_recipes_api}
\alias{add_custom_steps}
\alias{update_custom_steps}
\alias{remove_custom_steps}
\title{Add, update, or remove recipe recipes from a \code{synth_spec} object}
\usage{
add_custom_steps(synth_spec, ...)

update_custom_steps(synth_spec, ...)

remove_custom_steps(synth_spec)
}
\arguments{
\item{synth_spec}{A \code{synth_spec} object}

\item{...}{Optional named arguments mapping variables to lists of
\code{recipe::recipe_} function(s) from \code{library(recipes)}.}
}
\value{
A new \code{synth_spec} object.
}
\description{
Add, update, or remove recipe recipes from a \code{synth_spec} object
}
