% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synth_spec.R
\name{synth_spec_tuner_api}
\alias{synth_spec_tuner_api}
\alias{add_custom_tuners}
\alias{update_custom_tuners}
\alias{remove_custom_tuners}
\title{Add, update, or remove tuners from a \code{synth_spec} object}
\usage{
add_custom_tuners(synth_spec, ...)

update_custom_tuners(synth_spec, ...)

remove_custom_tuners(synth_spec)
}
\arguments{
\item{synth_spec}{A \code{synth_spec} object}

\item{...}{Optional named lists with two elements, \code{vars} and \code{tuner},
mapping variable names to tuners.}
}
\value{
A new \code{synth_spec} object.
}
\description{
Add, update, or remove tuners from a \code{synth_spec} object
}
