% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesize_j.R
\name{synthesize_j}
\alias{synthesize_j}
\title{synthesize the jth variable}
\usage{
synthesize_j(
  conf_data,
  synth_data,
  col_schema,
  model,
  recipe,
  sampler,
  noise,
  tuner,
  extractor,
  constraints,
  invert_transformations,
  p = function() {
     NULL
 }
)
}
\arguments{
\item{conf_data}{A data frame for estimating the predictive model}

\item{synth_data}{A data frame of non-sensitive predictors or synthesized
predictors}

\item{col_schema}{A list of schematic information about each variable.}

\item{model}{A \code{model_spec} or a list of \code{model_spec}s from \code{library(parsnip)}}

\item{recipe}{A recipe from \code{library(recipes)}.}

\item{sampler}{A method used to predict values from estimated models.}

\item{noise}{A \code{noise} object}

\item{tuner}{A \code{tuner} object for hyperparameter tuning. Hyperparameter
tuning is skipped if \code{tuner = NULL}.}

\item{extractor}{A method for extracting workflows or extracts from
workflows. \code{NULL} will extract no information. \code{"workflow"} will extract the
entire workflow. \verb{extract_*()} will use extract functions from
\code{library(workflows)} to extract information.}

\item{constraints}{A list of \code{constraints_df} and \code{max_z} values}

\item{invert_transformations}{A Boolean for if outcome variable
transformations should be inverted during synthesis}

\item{p}{A function for tracking progress created with progressr}
}
\value{
A list with an estimated model and a data frame of predictions
}
\description{
synthesize the jth variable
}
