% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_synthesis.R
\name{tune_synthesis}
\alias{tune_synthesis}
\title{Generate syntheses from multiple \code{presynth} objects.}
\usage{
tune_synthesis(
  presynths,
  postprocessing_func,
  metadata_func = NULL,
  simplify_post = FALSE,
  seed = NULL
)
}
\arguments{
\item{presynths}{A list of \code{presynth} objects}

\item{postprocessing_func}{A function with arguments "synth_id", "synth_name",
and "postsynth" that performs any desired postprocessing operations, like writing}

\item{metadata_func}{An optional function with argument "presynth" that extracts
specified information from each \code{presynth} object and returns a list. Each list
element becomes an additional column in the output metadata.}

\item{simplify_post}{Boolean that, if true, expects \code{postprocessing_func} to
return a list corresponding to the row of the output dataframe (one per synthesis).}

\item{seed}{A RNG seed to pass to \code{set.seed()}}
}
\value{
A \code{post_tunesynth} object.
}
\description{
Generate syntheses from multiple \code{presynth} objects.
}
