% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{validate_constraints}
\alias{validate_constraints}
\title{Validate constraints}
\usage{
validate_constraints(roadmap)
}
\arguments{
\item{roadmap}{A \code{roadmap} object}
}
\description{
Validate constraints
}
