% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roadmap.R
\name{validate_roadmap}
\alias{validate_roadmap}
\title{Roadmap validator}
\usage{
validate_roadmap(roadmap)
}
\arguments{
\item{roadmap}{A \code{roadmap} object.}
}
\description{
Ensures internal and external consistency between a \code{roadmap} and its constituent
objects, raising errors if the \code{roadmap} cannot be used for syntheses. This
function is lazily evaluated immediately prior to synthesis.
}
