% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_method.R
\name{validate_start_method}
\alias{validate_start_method}
\title{Validate a \code{start_method}}
\usage{
validate_start_method(roadmap)
}
\arguments{
\item{roadmap}{A \code{roadmap} object.}
}
\description{
Validate a \code{start_method}
}
