% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visit_sequence.R
\name{validate_visit_sequence}
\alias{validate_visit_sequence}
\title{Validate \code{visit_sequence}}
\usage{
validate_visit_sequence(roadmap)
}
\arguments{
\item{roadmap}{A \code{roadmap} object.}
}
\description{
Validate \code{visit_sequence}
}
