% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visit_sequence_api.R
\name{visit_sequence_api}
\alias{visit_sequence_api}
\alias{add_visit_sequence}
\alias{update_visit_sequence}
\alias{reset_visit_sequence}
\title{Add or reset a \code{visit_sequence} object within an existing \code{roadmap}.}
\usage{
add_visit_sequence(roadmap, visit_sequence)

update_visit_sequence(roadmap, ...)

reset_visit_sequence(roadmap)
}
\arguments{
\item{roadmap}{A \code{roadmap} object}

\item{visit_sequence}{A \code{visit_sequence} object.}

\item{...}{Optional additional parameters.}
}
\value{
A new \code{roadmap} object.
}
\description{
Add or reset a \code{visit_sequence} object within an existing \code{roadmap}.
}
\examples{

roadmap <- roadmap(
  conf_data = acs_conf_nw, 
  start_data = acs_start_nw
) 

roadmap |>
  update_visit_sequence(
    weight_var = wgt,
    synthesize_weight = TRUE
  )


roadmap <- roadmap(
  conf_data = acs_conf_nw, 
  start_data = acs_start_nw
) 

roadmap <- roadmap |>
  update_visit_sequence(
    weight_var = wgt,
    synthesize_weight = TRUE
  )

reset_visit_sequence(roadmap = roadmap)

}
