% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yeo_johnson.R
\name{yeo_johnson}
\alias{yeo_johnson}
\title{Perform a Yeo-Johnson transformation. Ported from library(VGAM)}
\usage{
yeo_johnson(
  y,
  lambda,
  derivative = 0,
  epsilon = sqrt(.Machine$double.eps),
  inverse = FALSE
)
}
\arguments{
\item{y}{Numeric, a vector or matrix.}

\item{lambda}{Numeric. It is recycled to the same length as y if necessary.}

\item{derivative}{Non-negative integer. The default is the ordinary function evaluation, otherwise the derivative with respect to lambda.}

\item{epsilon}{Numeric and positive value. The tolerance given to values of lambda when comparing it to 0 or 2.}

\item{inverse}{Logical. Return the inverse transformation?}
}
\value{
A transformed numeric vector
}
\description{
Perform a Yeo-Johnson transformation. Ported from library(VGAM)
}
