% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_noise.R
\name{construct_noise}
\alias{construct_noise}
\title{Construct a list of noise objects for synthesis}
\usage{
construct_noise(
  roadmap,
  default_regression_noise = NULL,
  default_classification_noise = NULL,
  custom_noise = NULL
)
}
\arguments{
\item{roadmap}{A roadmap object}

\item{default_regression_noise}{A noise function for regression models}

\item{default_classification_noise}{A noise function for classification models}

\item{custom_noise}{A formatted list of noise functions}
}
\value{
A named list of noise
}
\description{
Construct a list of noise objects for synthesis
}
