% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_recipes.R
\name{construct_recipes}
\alias{construct_recipes}
\title{Construct a sequence of model recipes for sequential synthesis}
\usage{
construct_recipes(
  roadmap,
  default_regression_steps = NULL,
  default_classification_steps = NULL,
  custom_steps = NULL
)
}
\arguments{
\item{roadmap}{A roadmap object}

\item{default_regression_steps}{A list containing one or more recipes::step_*()}

\item{default_classification_steps}{A list containing one or more recipes::step_*()}

\item{custom_steps}{A list of lists containing one or more recipes::step_*()}
}
\value{
A list of formulas
}
\description{
Construct a sequence of model recipes for sequential synthesis
}
