% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_samplers.R
\name{construct_samplers}
\alias{construct_samplers}
\title{Construct a list of samplers for synthesis}
\usage{
construct_samplers(
  roadmap,
  default_regression_sampler = NULL,
  default_classification_sampler = NULL,
  custom_samplers = NULL
)
}
\arguments{
\item{roadmap}{A roadmap object}

\item{default_regression_sampler}{A sampler function for regression models}

\item{default_classification_sampler}{A sampler function for classification models}

\item{custom_samplers}{A formatted list of sampler functions}
}
\value{
A named list of samplers
}
\description{
Construct a list of samplers for synthesis
}
\examples{

roadmap <- roadmap(
  conf_data = acs_conf_nw,
  start_data = acs_start_nw
)

construct_samplers(
  roadmap = roadmap, 
  default_regression_sampler = sample_rpart, 
  default_classification_sampler = sample_rpart
)

}
