% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_tuners.R
\name{construct_tuners}
\alias{construct_tuners}
\title{Construct a list of tuning grids for hyperparameter tuning predictive models}
\usage{
construct_tuners(
  roadmap,
  default_regression_tuner = NULL,
  default_classification_tuner = NULL,
  custom_tuners = NULL
)
}
\arguments{
\item{roadmap}{A roadmap object}

\item{default_regression_tuner}{A tuner.}

\item{default_classification_tuner}{A tuner.}

\item{custom_tuners}{A formatted list of tuners.}
}
\value{
A named list of tuners
}
\description{
Construct a list of tuning grids for hyperparameter tuning predictive models
}
\examples{

roadmap <- roadmap(
  conf_data = acs_conf_nw,
  start_data = acs_start_nw
)

tuner_reg <- list(
  v = 3,
  grid = 3,
  metrics = yardstick::metric_set(yardstick::rmse)
)

tuner_cat <- list(
  v = 3,
  grid = 3,
  metrics = yardstick::metric_set(yardstick::roc_auc)
)

construct_tuners(
  roadmap = roadmap, 
  default_regression_tuner = tuner_reg, 
  default_classification_tuner = tuner_cat
)

}
