% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_level_to_na.R
\name{convert_level_to_na}
\alias{convert_level_to_na}
\title{Convert \code{"NA"} values to \code{NA} for categorical variables}
\usage{
convert_level_to_na(data)
}
\arguments{
\item{data}{A data frame or tibble}
}
\value{
A data frame or tibble with \code{"NA"} converted to \code{NA}
}
\description{
Convert \code{"NA"} values to \code{NA} for categorical variables
}
