% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synth_component_utils.R
\docType{data}
\name{.name_to_inspect}
\alias{.name_to_inspect}
\title{constant mapping between component names and inspections}
\format{
An object of class \code{list} of length 7.
}
\usage{
.name_to_inspect
}
\description{
constant mapping between component names and inspections
}
\keyword{datasets}
