% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enforce_schema.R
\name{enforce_schema}
\alias{enforce_schema}
\title{Enforce a \code{roadmap}'s \code{schema} on its existing data}
\usage{
enforce_schema(roadmap)
}
\arguments{
\item{roadmap}{A \code{roadmap} object}
}
\value{
A \code{roadmap} object with modified \code{conf_data}, \code{start_data}, and \code{schema}
information.
}
\description{
Enforce a \code{roadmap}'s \code{schema} on its existing data
}
\examples{
  
roadmap <- roadmap(conf_data = acs_conf, start_data = acs_start) |>
  update_schema(na_numeric_to_ind = TRUE)

enforce_schema(roadmap)

}
