% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invert.R
\name{invert.step_log}
\alias{invert.step_log}
\title{Invert a log transformation}
\usage{
\method{invert}{step_log}(object, predictions, ...)
}
\arguments{
\item{object}{A recipe after fitting a model}

\item{predictions}{A data frame with .pred}

\item{...}{Other arguments}
}
\description{
Invert a log transformation
}
