% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invert.R
\name{invert.step_YeoJohnson}
\alias{invert.step_YeoJohnson}
\title{Invert a Yeo-Johnson transformation}
\usage{
\method{invert}{step_YeoJohnson}(object, predictions, ...)
}
\arguments{
\item{object}{A recipe after fitting a model}

\item{predictions}{A data frame with .pred}

\item{...}{Other arguments}
}
\value{
A tibble with Yeo_johnson transformation inverted for .pred
}
\description{
Invert a Yeo-Johnson transformation
}
