% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset_postsynth.R
\name{postsynth_to_roadmap}
\alias{postsynth_to_roadmap}
\title{Convert a incompletely synthesized \code{postsynth} into a new \code{roadmap} using
the same \code{roadmap} settings from the incomplete synthesis.}
\usage{
postsynth_to_roadmap(postsynth)
}
\arguments{
\item{postsynth}{A \code{postsynth} object generated with \code{keep_workflows == TRUE}}
}
\value{
A \code{roadmap} object.
}
\description{
Convert a incompletely synthesized \code{postsynth} into a new \code{roadmap} using
the same \code{roadmap} settings from the incomplete synthesis.
}
