% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset_postsynth.R
\name{postsynth_to_synth_spec}
\alias{postsynth_to_synth_spec}
\title{Convert a incompletely synthesized \code{postsynth} into a new \code{synth_spec} using
the same \code{synth_spec} settings from the incomplete synthesis.}
\usage{
postsynth_to_synth_spec(postsynth)
}
\arguments{
\item{postsynth}{A \code{postsynth} object generated with \code{keep_workflows == TRUE}}
}
\value{
A \code{synth_spec} object.
}
\description{
Convert a incompletely synthesized \code{postsynth} into a new \code{synth_spec} using
the same \code{synth_spec} settings from the incomplete synthesis.
}
