% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presynth.R
\name{presynth}
\alias{presynth}
\title{Create a presynth object}
\usage{
presynth(roadmap, synth_spec)
}
\arguments{
\item{roadmap}{A \code{roadmap} object from \code{roadmap()}.}

\item{synth_spec}{A \code{synth_spec} object from \code{synth_spec()}.}
}
\value{
A \code{presynth} object.
}
\description{
Create a presynth object
}
