% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postsynth.R
\name{print.postsynth}
\alias{print.postsynth}
\title{Print the postsynth object to the console with formatting}
\usage{
\method{print}{postsynth}(x, ...)
}
\arguments{
\item{x}{A \code{postsynth} object}

\item{...}{further arguments passed to or from other methods (not currently
used).}
}
\description{
Print the postsynth object to the console with formatting
}
