% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicates.R
\name{print.replicates}
\alias{print.replicates}
\title{Print the replicates object to the console with formatting}
\usage{
\method{print}{replicates}(x, ...)
}
\arguments{
\item{x}{A \code{replicates} object}

\item{...}{further arguments passed to or from other methods (not currently
used).}
}
\description{
Print the replicates object to the console with formatting
}
\examples{

replicates <- replicates(
  start_data_replicates = 2,
  model_sample_replicates = 2,
  end_to_end_replicates = 2
)
  
print(replicates)
  
}
