% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{print.schema}
\alias{print.schema}
\title{Print the schema object to the console with formatting}
\usage{
\method{print}{schema}(x, ...)
}
\arguments{
\item{x}{A \code{schema} object}

\item{...}{further arguments passed to or from other methods (not currently
used).}
}
\value{
Invisibly returns the input \code{schema} object.
}
\description{
Print the schema object to the console with formatting
}
\examples{

# default inferred schema
schema1 <- schema(
  conf_data = acs_conf_nw,
  start_data = acs_start_nw
)

print(schema1)

}
