% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_lm.R
\name{sample_lm}
\alias{sample_lm}
\title{Sample the conditional distribution created by a linear model}
\usage{
sample_lm(model, new_data, conf_data)
}
\arguments{
\item{model}{A "model_fit" object created by parsnip::linear_reg()}

\item{new_data}{A data frame with predictors}

\item{conf_data}{A data frame with original confidential predictors}
}
\value{
A numeric vector of predictions
}
\description{
Sample the conditional distribution created by a linear model
}
