% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_ranger.R
\name{sample_ranger}
\alias{sample_ranger}
\title{Sample the conditional distribution created by a ranger rf model}
\usage{
sample_ranger(model, new_data, conf_data)
}
\arguments{
\item{model}{A "model_fit" object created by parsnip::ranger()}

\item{new_data}{A data frame with predictors}

\item{conf_data}{A data frame with original confidential predictors}
}
\value{
A numeric vector of predictions
}
\description{
Sample the conditional distribution created by a ranger rf model
}
