% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{schema_api}
\alias{schema_api}
\alias{add_schema}
\alias{update_schema}
\alias{reset_schema}
\title{Add, update, or reset a \code{schema} object within an existing \code{roadmap}.}
\usage{
add_schema(roadmap, schema)

update_schema(roadmap, ...)

reset_schema(roadmap)
}
\arguments{
\item{roadmap}{A \code{roadmap} object}

\item{schema}{A \code{schema} object.}

\item{...}{Optional named parameters passed to \code{schema()}.}
}
\value{
A new \code{roadmap} object.
}
\description{
Add, update, or reset a \code{schema} object within an existing \code{roadmap}.
}
