% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesize.R
\name{synthesize}
\alias{synthesize}
\title{Synthesize a data set}
\usage{
synthesize(presynth, progress = FALSE)
}
\arguments{
\item{presynth}{A \code{presynth} object created by \code{presynth()}.}

\item{progress}{A single logical. Should a progress be displayed?}
}
\value{
A postsynth object.
}
\description{
Synthesize a data set
}
