% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presynth.R
\name{update_presynth}
\alias{update_presynth}
\title{Update \code{presynth} object}
\usage{
update_presynth(presynth, roadmap = NULL, synth_spec = NULL)
}
\arguments{
\item{presynth}{A \code{presynth} object}

\item{roadmap}{An optional \code{roadmap} object}

\item{synth_spec}{An optional \code{synth_spec} object}
}
\value{
A \code{presynth} object.
}
\description{
Update \code{presynth} object
}
