% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synth_spec.R
\name{update_synth_spec}
\alias{update_synth_spec}
\title{Tidy API calls ----------------------------------------------------------}
\usage{
update_synth_spec(synth_spec, ...)
}
\arguments{
\item{synth_spec}{A \code{synth_spec} object}

\item{...}{Optional named keywords in \code{synth_spec}, with the exception of
any \verb{custom_*}arguments}
}
\value{
A \code{synth_spec}
}
\description{
Update non-custom \code{synth_spec} arguments
}
